cleanup(ε.4): extract decode_block/fill_symbol_spectra.rs (per-symbol DFT stage)
Fourth slice of the `decode_block.rs` per-stage split. Moves the
per-candidate symbol-spectrum fill stage into a sibling submodule:

- `SymMask` enum (Costas-first symbol-selection mask used by the
  Pass-1 sync_quality gate)
- `symbol_spectra_direct()` — public `f32` entry that allocates +
  fills the 79×8 cs box
- `fill_symbol_spectra()` — host (`fft-rustfft`, via cd0) and
  embedded (rectangular per-tone DFT, non-fixed-point) variants
- `fill_symbol_spectra_via_cd0()` — private cd0-based 32-pt FFT
  implementation (`ft8b.f90:154-161` mirror)
- `fill_symbol_spectra_generic<Sc, S>()` — generic-scalar entry
  used by the embedded fixed-point Q11i16 LLR path
- `fill_symbol_spectra_into()` / `_generic` — caller-supplied basis
  variants (fixed-point only)
- `symbol_spectra_direct_into()` — embedded one-shot direct DFT
- `BASIS_SCRATCH_LEN` — `NTONES * NSPS` constant the embedded apps
  size their per-thread basis scratch against
- `sym_in_mask()` — private mask predicate

External-caller path preserved: parent re-exports the public
entries via `pub use fill_symbol_spectra::*` (split into
fixed-point and non-fixed-point re-export groups so the cfg-gated
items don't leak across builds). The host `super::decode::*` path
keeps reaching `SymMask` + `fill_symbol_spectra` via the same
`mfsk_core::ft8::decode_block::*` path; `mfsk-ffi-ft8` keeps
reaching `BASIS_SCRATCH_LEN`.

Parent cleanup: `Complex`, `COSTAS_POS`, `NN` imports dropped from
the top-level `use` block (only fill_symbol_spectra used them).

Acceptance:
- `cargo check` clean on all 4 feature cells
- FT8 qso3_busy golden tests pass (WSJT-X AP-off floor, JTDX
  recall floor, AP-on strict-superset, AP-on subtract diag)
- 7 `ft8::decode_block::tests::*` lib tests pass
- `cargo clippy --features full --lib` clean, `cargo fmt --check` clean
- Diff: decode_block.rs −575 / +12, new fill_symbol_spectra.rs +598

`decode_block.rs` is now down to 1,949 lines (started at 3,517 —
45 % reduction). Remaining ε slices: process_candidates (ε.5),
osd_strategy (ε.6).

Stacks on #79 (ε.3). Rebases to main when #79 lands.
